Mit Gates die Zuweisung von Tags autorisieren

+++ app/Http/Controllers/hobbyTagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Hobby;
 use Illuminate\Http\Request;
 use App\Tag;
+use Illuminate\Support\Facades\Gate;
 
 class hobbyTagController extends Controller
 {
@@ -30,6 +31,11 @@ class hobbyTagController extends Controller
     public function attachTag($hobby_id, $tag_id){
 
         $hobby = Hobby::find($hobby_id);
+
+        if (Gate::denies('connect_hobbyTag', $hobby)) {
+            abort(403, 'Das Hobby ist nicht von Dir.');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->attach($tag_id);
 
@@ -41,6 +47,11 @@ class hobbyTagController extends Controller
     public function detachTag($hobby_id, $tag_id){
 
         $hobby = Hobby::find($hobby_id);
+
+        if (Gate::denies('connect_hobbyTag', $hobby)) {
+            abort(403, 'Das Hobby ist nicht von Dir.');
+        }
+
         $tag = Tag::find($tag_id);
         $hobby->tags()->detach($tag_id);
